Use SharedDriverConformanceTest for ReplicaSet
