+++ bosk-core/src/test/java/works/bosk/drivers/ReplicaSetConformanceTest.java
@@ -4,13 +4,13 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import works.bosk.Bosk;
 import works.bosk.BoskConfig;
-import works.bosk.testing.drivers.DriverConformanceTest;
+import works.bosk.testing.drivers.SharedDriverConformanceTest;
 import works.bosk.testing.drivers.state.TestEntity;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static works.bosk.testing.BoskTestUtils.boskName;
 
-class ReplicaSetConformanceTest extends DriverConformanceTest {
+class ReplicaSetConformanceTest extends SharedDriverConformanceTest {
 	Bosk<TestEntity> replicaBosk;
 
 	@BeforeEach
