+++ bosk-testing/src/main/java/works/bosk/testing/drivers/AbstractDriverTest.java
@@ -59,7 +59,7 @@ public abstract class AbstractDriverTest {
 	public enum Scenario {
 		NO_TENANTS(TenancyModel.NONE, Tenant.NONE),
 		FIXED_TENANT(new Fixed(TENANT1), Tenant.setTo(TENANT1)),
-		TRANSIENT(TenancyModel.TRANSIENT, Tenant.setTo(TENANT1))
+		TRANSIENT_TENANT(TenancyModel.TRANSIENT, Tenant.setTo(TENANT1))
 		;
 
 		public final TenancyModel tenancyModel;
